+++ src/components/CenteredPage/CenteredPage.scss
