+++ src/components/LockButton/LockButton.scss
