+++ src/pages/EntryPoint/EntryPoint.js
