+++ src/pages/EntryPoint/EntryPoint.scss
