+++ src/pages/Init/Init.scss
