+++ src/pages/NewEntry/NewEntry.js
