+++ src/pages/NewEntry/NewEntry.scss
