+++ src/pages/NewEntry/NewEntryContainer.js
