+++ src/pages/NewEntry/index.js
